+++ src/main/java/pt/codeforge/toolertools/xml/DocConstants.java
@@ -1,0 +1,12 @@
+package pt.codeforge.toolertools.xml;
+
+class DocConstants {
+
+    private DocConstants() {
+    }
+
+    protected static final String NAMESPACES = "http://xml.org/sax/features/namespaces";
+    protected static final String VALIDATION = "http://xml.org/sax/features/validation";
+    protected static final String LOAD_DTD_GRAMMAR = "http://apache.org/xml/features/nonvalidating/load-dtd-grammar";
+    protected static final String LOAD_EXTERNAL_DTD = "http://apache.org/xml/features/nonvalidating/load-external-dtd";
+}